requirements.txt
PyYAML==5.4.1
watchdog==1.0.2
